requirements.txt
aniso8601==9.0.1
attrs==21.4.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
Flask==2.0.3
Flask-PyMongo==2.3.0
flask-restplus==0.13.0
flask-restx==0.5.1
itsdangerous==2.1.0
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.1.0
passlib==1.7.4
pycparser==2.21
PyJWT==2.3.0
pymongo==4.0.1
pyrsistent==0.18.1
pytz==2021.3
six==1.16.0
typing_extensions==4.1.1
Werkzeug==2.0.3